feat: observable cookies

+++ src/cookie.wrapper.ts
@@ -1,18 +1,24 @@
 import Cookie from './cookie';
+
+export interface ICookieKeys {
+    [a: string]: Cookie
+}
 /**
  * Wraps Cookie sessions into a decent-to-use class for TypeScript projects.
  * 
  * Allows for most usual operations done over cookies, ie:
- *  - get a Cookie - getKey(<key:string>)
- *  - set a Cookie - setKey(<key:string>, <value:string>)
- *  - remove a Cookie - removeKey(<key:string>)
- *  - list all Cookie keys - queryKeys()
+ *  - get a Cookie - get(<key:string>)
+ *  - set a Cookie - set(<key:string>, <value:string>)
+ *  - remove a Cookie - remove(<key:string>)
+ *  - list all Cookie keys - query()
  */
 export default class CookieWrapper {
-    private keys: any = {};
+    private keys: ICookieKeys = {};
+    private domain: string;
 
-    constructor(private domain: string = "") {
-        this.parseKeys();
+    constructor(domain: string = "") {
+        this.domain = domain;
+        this.parse();
     }
 
     /**
@@ -20,8 +26,8 @@ export default class CookieWrapper {
      * @param key
      * @returns {Cookie}
      */
-    public getKey(key: string): Cookie {
-        this.parseKeys();
+    public get(key: string): Cookie {
+        this.parse();
         if (!this.keys[key]) return undefined;
         return this.keys[key];
     }
@@ -33,10 +39,9 @@ export default class CookieWrapper {
      * @param expiration
      * @returns {Cookie}
      */
-    public setKey(key: string, value: any, expiration: string = ''): Cookie {
-        let cookie: Cookie = new Cookie(key, value, this.domain, expiration);
-        this.keys[key] = cookie;
-        return cookie;
+    public set(key: string, value: string, expiration: string = ''): Cookie {
+        this.keys[key] = new Cookie(key, value, this.domain, expiration);
+        return this.keys[key];
     }
 
     /**
@@ -44,8 +49,9 @@ export default class CookieWrapper {
      * @deprecated: use the cookie object itself after using getKey
      * @param key
      */
-    public removeKey(key: string) {
-        let cookie = this.keys[key];
+    public remove(key: string) {
+        const cookie = this.keys[key];
+        if (!cookie) return;
         cookie.delete();
     }
 
@@ -53,9 +59,9 @@ export default class CookieWrapper {
      * Returns a list of all the Cookie keys
      * @returns {Array<string>}
      */
-    public queryKeys(): Array<string> {
-        this.parseKeys();
-        if (!Object.keys(this.keys)) return [];
+    public query(): Array<string> {
+        this.parse();
+        if (!Object.keys(this.keys).length) return [];
         return Object.keys(this.keys);
     }
 
@@ -65,11 +71,11 @@ export default class CookieWrapper {
      * @param days
      * @returns {Cookie}
      */
-    public setExpiration(key: string, days: number): Cookie {
-        this.parseKeys();
-        let expiration = new Date();
+    public expireIn(key: string, days: number): Cookie {
+        this.parse();
+        const expiration = new Date();
         expiration.setDate(expiration.getDate() + days);
-        let cookie = this.getKey(key);
+        const cookie = this.get(key);
         cookie.expires = expiration.toISOString();
         cookie.save();
         return cookie;
@@ -79,21 +85,21 @@ export default class CookieWrapper {
      * Populates this.keys as a Javascript Object indexing all Cookie objects currently in the
      * browser session
      */
-    private parseKeys() {
+    private parse() {
         let crumbs = document.cookie.split("; ");
         if (!crumbs[crumbs.length-1]) {
             crumbs.pop();
         }
         let keys = JSON.parse("{" +
             crumbs.map(crumb => {
-                let crumbArr = crumb.split(/=/);
+                const crumbArr = crumb.split(/=/);
                 return `"${crumbArr[0]}": "${crumbArr[1]}"`;
             }).join(",") +
         "}");
         this.keys = {};
         Object.keys(keys).forEach(key => {
             this.keys[key] = new Cookie(
-                key, keys[key].value, keys[key].domain, keys[key].expires, keys[key].path
+                key, keys[key], keys[key].domain, keys[key].expires, keys[key].path
             );
         });
     }
